// src/app/lib/types.ts
export interface glAccountType {
  idGlAccount?: number;
  uniqueId?: string;
  glAccount: string;
  description?: string;
  groupDetail: string;
  groupGl: string;
  sap: boolean;
  active: boolean;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string | null;
}

export interface mGlAccountType {
  idGlAccount: number;
  glAccount: string;
  groupGL: string;
  groupDetail: string;
}
export interface mCostCenterType {
  idCostCenter: number;
  costCenter: string;
  bidang: string;
  dinas: string;
}
export interface availableType {
  available: number;
  mGlAccount: mGlAccountType;
  mCostCenter: mCostCenterType;
}
export interface selectType {
  id: string | number;
  value: string;
}